Add files via upload
Upload the Code to Arduino.
Connect Pin 8 of Arduino to APM Input(Channel 5, if it is the Flight Mode channel). Connect Vcc and Gnd frm arduino to APM
Power up the Quadcopter WITHOUT PROPS.
Use Telemetry Module, connect to Mission Planner, Go To Flight Modes to see how it is performing.